+++ tests/Silverback.Integration.Kafka.Tests/Messaging/KafkaConsumerEndpointTests.cs
@@ -2,6 +2,8 @@
 // This code is licensed under MIT license (see LICENSE file for details)
 
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using Confluent.Kafka;
 using FluentAssertions;
 using Silverback.Messaging;
@@ -37,8 +39,8 @@ namespace Silverback.Tests.Integration.Kafka.Messaging
         {
             var endpoint = new KafkaConsumerEndpoint(new TopicPartition("topic", 2));
 
-            endpoint.Name.Should().Be("topic[2]");
-            endpoint.Names.Should().BeEquivalentTo("topic[2]");
+            endpoint.Name.Should().Be("topic");
+            endpoint.Names.Should().BeEquivalentTo("topic");
             endpoint.TopicPartitions.Should().BeEquivalentTo(
                 new TopicPartitionOffset("topic", 2, Offset.Unset));
         }
@@ -52,8 +54,8 @@ namespace Silverback.Tests.Integration.Kafka.Messaging
                 new TopicPartition("topic2", 2),
                 new TopicPartition("topic2", 3));
 
-            endpoint.Name.Should().Be("[topic1[0],topic1[1],topic2[2],topic2[3]]");
-            endpoint.Names.Should().BeEquivalentTo("topic1[0]", "topic1[1]", "topic2[2]", "topic2[3]");
+            endpoint.Name.Should().Be("[topic1,topic2]");
+            endpoint.Names.Should().BeEquivalentTo("topic1", "topic2");
             endpoint.TopicPartitions.Should().BeEquivalentTo(
                 new TopicPartitionOffset("topic1", 0, Offset.Unset),
                 new TopicPartitionOffset("topic1", 1, Offset.Unset),
@@ -70,13 +72,77 @@ namespace Silverback.Tests.Integration.Kafka.Messaging
                 new TopicPartitionOffset("topic2", 2, 42),
                 new TopicPartitionOffset("topic2", 3, Offset.Unset));
 
-            endpoint.Name.Should().Be("[topic1[0],topic1[1],topic2[2],topic2[3]]");
-            endpoint.Names.Should().BeEquivalentTo("topic1[0]", "topic1[1]", "topic2[2]", "topic2[3]");
+            endpoint.Name.Should().Be("[topic1,topic2]");
+            endpoint.Names.Should().BeEquivalentTo("topic1", "topic2");
             endpoint.TopicPartitions.Should().BeEquivalentTo(
                 new TopicPartitionOffset("topic1", 0, Offset.Beginning),
                 new TopicPartitionOffset("topic1", 1, Offset.End),
                 new TopicPartitionOffset("topic2", 2, 42),
                 new TopicPartitionOffset("topic2", 3, Offset.Unset));
+        }
+
+        [Fact]
+        public void Constructor_SingleTopicAndPartitionResolver_TopicsAndResolverSet()
+        {
+            var endpoint = new KafkaConsumerEndpoint(
+                "topic1",
+                partitions => partitions);
+
+            endpoint.Name.Should().Be("topic1");
+            endpoint.Names.Should().BeEquivalentTo("topic1");
+            endpoint.TopicPartitions.Should().BeNull();
+            endpoint.TopicPartitionsResolver.Should().NotBeNull();
+            endpoint.TopicPartitionsResolver.Should()
+                .BeOfType<Func<IReadOnlyCollection<TopicPartition>, IEnumerable<TopicPartitionOffset>>>();
+        }
+
+        [Fact]
+        public void Constructor_MultipleTopicsAndPartitionResolver_TopicsAndResolverSet()
+        {
+            var endpoint = new KafkaConsumerEndpoint(
+                new[] { "topic1", "topic2" },
+                partitions => partitions);
+
+            endpoint.Name.Should().Be("[topic1,topic2]");
+            endpoint.Names.Should().BeEquivalentTo("topic1", "topic2");
+            endpoint.TopicPartitions.Should().BeNull();
+            endpoint.TopicPartitionsResolver.Should().NotBeNull();
+            endpoint.TopicPartitionsResolver.Should()
+                .BeOfType<Func<IReadOnlyCollection<TopicPartition>, IEnumerable<TopicPartitionOffset>>>();
+        }
+
+        [Fact]
+        public void Constructor_SingleTopicAndPartitionResolverWithOffsets_TopicsAndResolverSet()
+        {
+            Func<IReadOnlyCollection<TopicPartition>, IEnumerable<TopicPartitionOffset>> resolver =
+                partitions => partitions.Select(
+                    partition => new TopicPartitionOffset(partition, Offset.Beginning));
+
+            var endpoint = new KafkaConsumerEndpoint(
+                "topic1",
+                resolver);
+
+            endpoint.Name.Should().Be("topic1");
+            endpoint.Names.Should().BeEquivalentTo("topic1");
+            endpoint.TopicPartitions.Should().BeNull();
+            endpoint.TopicPartitionsResolver.Should().Be(resolver);
+        }
+
+        [Fact]
+        public void Constructor_MultipleTopicsAndPartitionResolverWithOffsets_TopicsAndResolverSet()
+        {
+            Func<IReadOnlyCollection<TopicPartition>, IEnumerable<TopicPartitionOffset>> resolver =
+                partitions => partitions.Select(
+                    partition => new TopicPartitionOffset(partition, Offset.Beginning));
+
+            var endpoint = new KafkaConsumerEndpoint(
+                new[] { "topic1", "topic2" },
+                resolver);
+
+            endpoint.Name.Should().Be("[topic1,topic2]");
+            endpoint.Names.Should().BeEquivalentTo("topic1", "topic2");
+            endpoint.TopicPartitions.Should().BeNull();
+            endpoint.TopicPartitionsResolver.Should().Be(resolver);
         }
 
         [Fact]