+++ docs/releases.md
@@ -21,6 +21,7 @@ uid: releases
 
 * Prevent possible race condition causing messages to be skipped when a `RetryPolicy` kicks in for messages from multiple Kafka partitions simultaneously
 * Prevent `ObjectDisposedException` to be thrown when Kafka events (e.g. statistics) are fired during the application shutdown
+* Prevent `ObjectDisposedException` to be thrown when `Consumer.Dispose` is called multiple times
 
 ## [3.0.1](https://github.com/BEagle1984/silverback/releases/tag/v3.0.1)